requirements.txt
annotated-types==0.7.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
dotenv==0.9.9
filelock==3.20.0
Flask==3.1.2
fsspec==2025.9.0
google-ai-generativelanguage==0.6.15
google-api-core==2.26.0
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.71.2
hf-xet==1.1.10
httplib2==0.31.0
huggingface-hub==0.35.3
idna==3.10
inquirerpy==0.3.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
pfzy==0.3.4
pillow==11.3.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.0
pydantic_core==2.41.1
pyparsing==3.2.5
python-dotenv==1.1.1
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.14
Werkzeug==3.1.3
